v2/requirements.txt
flask
google-cloud-pubsub
opencv-python
pytesseract
easyocr
matplotlib
pytubefix
